backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic[email]==2.5.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
groq==0.4.2
httpx==0.26.0
python-dotenv==1.0.0
